chore(dev-cli): update command registration
